bugfix, cleanups
- fixed bug in system_checkup, it was using page zero locations it
shouldn't have.
- added build script for Linux (I'm actually using WSL).
- fixed some label names and added diagnostics to the build script to
show remaining free memory at build time.

Switched to a newer ATasm version, for the first time since I started
using it in 2005-04-24. Now switched from v1.07 to this commit (v1.27
plus a bug fix):
https://github.com/CycoPH/atasm/commit/ff3fb5ef4082b9a0179ddc4baa8d7824b66a035a
The reason for the upgrade is that 1.07 did not show equates (only
symbols) in the -l output, so the diagnostics added to the build script
couldn't work.