Fixed logic for displaying colored space characters.
Previously, when displaying colored characters and a space bar is
emitted, we preferred to skip changing the color at that position, to
prevent spoiling the color of an adjacent colored character, unless we
are in underline or inverse mode, in which case we're actually
displaying a character so we must color it.
Added two more conditions: if we are overwriting a previously displayed
character, or if the color was set as a background color, do color the
new space.