Fixed a corner case when scrolling the color underlay.
If forcing the color underlay to scroll with extended command 5 or 6,
with bitmap scroll off and rotation off, the new line will be colored
with the currently set background color as part of the scroll process.
The previous method was to first put zeros in the color, then fill the
colors of the new line in a very cumbersome and slow way, causing brief
visual glitches.