Changed PM layout and added reset_pms subroutine.
Since we want to consider the missiles as the fifth player for private
extensions 11-15, changed tables at initial_program_entry?tb_hi and
tb_lo such that entry 'n' will point to data of player 'n' (instead of
missiles being entry 0, reflecting their natural positioning in memory).
This imposed a rearrangement of the boldface underlay, so now the
missiles are on the right side of the screen instead of the left.
Also added reset_pms, to be called whenever the terminal is reset in
order to reinitialize the PM positions, sizes, and priority.